Refactor QVariantList to qtcore::qlist::qvariantlist
