+++ qttypes/src/qtcore/mod.rs
@@ -8,6 +8,6 @@ mod qurl;
 pub use self::primitives::qreal;
 pub use self::qbytearray::QByteArray;
 pub use self::qchar::UnicodeVersion;
-pub use self::qlist::QStringList;
+pub use self::qlist::{QStringList, QVariantList};
 pub use self::qstring::{NormalizationForm, QString};
 pub use self::qurl::QUrl;
