+++ qttypes/src/qtcore/qlist/mod.rs
@@ -1,3 +1,5 @@
 mod qstringlist;
+mod qvariantlist;
 
 pub use self::qstringlist::QStringList;
+pub use self::qvariantlist::QVariantList;
